@Groxx 2018-05-18
... huh.  `./versioned_go_build.sh: line 104: DEBUG: unbound variable`
bash version difference maybe?  can't reproduce.  fixed anyway.
and I wonder if grep is old or something on travis, it seems to be interpreting the regex wrong (version tags are being detected as non-versioned).  trying a fix and rerunning from what appear to be unrelated test failures.