@Groxx 2018-05-18
I'm thinking of making a bins folder, actually.  something that's not cleaned, so versioned-bins don't need to be redownloaded just because someone wanted to rebuild.